fix: preserve request_detail_id in transaction update and fix partial outstanding query
- Preserve request_detail_id when upserting transaction details so
  completeTransaction can correctly mark request details as 'realized'
- Update request_detail status in updateTransaction when draft → completed
- Send notification to request creator on realization via update flow
- Fix outstanding index diffInDays cast to int

+++ resources/views/transaction/outstanding/index.blade.php
@@ -96,7 +96,7 @@
                     $typeKey = $r->trans_code == 1 ? 'in' : 'out';
                     $typeLabel = $r->trans_code == 1 ? 'Masuk' : 'Keluar';
                     $typeBg = $r->trans_code == 1 ? 'bg-green-lt text-green' : 'bg-red-lt text-red';
-                    $days = now()->diffInDays($r->created_at);
+                    $days = (int) now()->diffInDays($r->created_at);
                 @endphp
                 <tr data-request-id="{{ $r->id }}">
                     <td><span class="badge {{ $typeBg }}">{{ $typeLabel }}</span></td>
@@ -177,7 +177,7 @@
                     $typeKey = $r->trans_code == 1 ? 'in' : 'out';
                     $typeLabel = $r->trans_code == 1 ? 'Masuk' : 'Keluar';
                     $typeBg = $r->trans_code == 1 ? 'bg-green-lt text-green' : 'bg-red-lt text-red';
-                    $daysApproved = $r->approved_at ? now()->diffInDays($r->approved_at) : 0;
+                    $daysApproved = $r->approved_at ? (int) now()->diffInDays($r->approved_at) : 0;
                 @endphp
                 <tr data-request-id="{{ $r->id }}">
                     <td><span class="badge {{ $typeBg }}">{{ $typeLabel }}</span></td>
